Implements user authentication UI
Adds Material UI components and styling to the login, registration, and home pages for improved user experience.

Introduces theme context and toggle for managing light/dark mode preferences and improves i18n.

The main goal is to provide UI elements and styling for logging in, registering, and basic home page functionality as well as providing a solid base upon which to build the entire application.